renaming types

+++ base/src/FalmCFDL1.h
@@ -8,21 +8,21 @@ namespace Falm {
 
 class L1Explicit {
 public:
-    double Re;
-    double ReI;
-    double dt;
+    REAL Re;
+    REAL ReI;
+    REAL dt;
 
-    L1Explicit(double _Re, double _dt): Re(_Re), ReI(1 / _Re), dt(_dt) {}
+    L1Explicit(REAL _Re, REAL _dt): Re(_Re), ReI(1 / _Re), dt(_dt) {}
 
     void L1Dev_Cartesian_FSCalcPseudoU(
-        Matrix<double> &u,
-        Matrix<double> &uu,
-        Matrix<double> &ua,
-        Matrix<double> &nut,
-        Matrix<double> &kx,
-        Matrix<double> &g,
-        Matrix<double> &jac,
-        Matrix<double> &ff,
+        Matrix<REAL> &u,
+        Matrix<REAL> &uu,
+        Matrix<REAL> &ua,
+        Matrix<REAL> &nut,
+        Matrix<REAL> &kx,
+        Matrix<REAL> &g,
+        Matrix<REAL> &jac,
+        Matrix<REAL> &ff,
         Mapper         &proc_domain,
         dim3            block_dim
     ) {
@@ -30,10 +30,10 @@ public:
         L0Dev_Cartesian_FSCalcPseudoU(u, uu, ua, nut, kx, g, jac, ff, proc_domain, map, block_dim);
     }
     void L1Dev_Cartesian_UtoCU(
-        Matrix<double> &u,
-        Matrix<double> &uc,
-        Matrix<double> &kx,
-        Matrix<double> &jac,
+        Matrix<REAL> &u,
+        Matrix<REAL> &uc,
+        Matrix<REAL> &kx,
+        Matrix<REAL> &jac,
         Mapper         &proc_domain,
         dim3            block_dim
     ) {
@@ -43,23 +43,23 @@ public:
 
 protected:
     void L0Dev_Cartesian_FSCalcPseudoU(
-        Matrix<double> &u,
-        Matrix<double> &uu,
-        Matrix<double> &ua,
-        Matrix<double> &nut,
-        Matrix<double> &kx,
-        Matrix<double> &g,
-        Matrix<double> &jac,
-        Matrix<double> &ff,
+        Matrix<REAL> &u,
+        Matrix<REAL> &uu,
+        Matrix<REAL> &ua,
+        Matrix<REAL> &nut,
+        Matrix<REAL> &kx,
+        Matrix<REAL> &g,
+        Matrix<REAL> &jac,
+        Matrix<REAL> &ff,
         Mapper         &proc_domain,
         Mapper         &map,
         dim3            block_dim
     );
     void L0Dev_Cartesian_UtoCU(
-        Matrix<double> &u,
-        Matrix<double> &uc,
-        Matrix<double> &kx,
-        Matrix<double> &jac,
+        Matrix<REAL> &u,
+        Matrix<REAL> &uc,
+        Matrix<REAL> &kx,
+        Matrix<REAL> &jac,
         Mapper         &proc_domain,
         Mapper         &map,
         dim3            block_dim
