+++ base/src/MVL2.h
@@ -6,27 +6,27 @@
 
 namespace Falm {
 
-static double L2Dev_DotProduct(Matrix<double> &a, Matrix<double> &b, Mapper &pdom, dim3 block_dim, CPMBase &cpm) {
+static REAL L2Dev_DotProduct(Matrix<REAL> &a, Matrix<REAL> &b, Mapper &pdom, dim3 block_dim, CPMBase &cpm) {
     Mapper map(pdom, Gd);
-    double r = L0Dev_DotProduct(a, b, pdom, map, block_dim);
+    REAL r = L0Dev_DotProduct(a, b, pdom, map, block_dim);
     if (cpm.size > 1) {
         CPML2_AllReduce(&r, 1, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
     }
     return r;
 }
 
-static double L2Dev_Norm2Sq(Matrix<double> &a, Mapper &pdom, dim3 block_dim, CPMBase &cpm) {
+static REAL L2Dev_Norm2Sq(Matrix<REAL> &a, Mapper &pdom, dim3 block_dim, CPMBase &cpm) {
     Mapper map(pdom, Gd);
-    double r = L0Dev_Norm2Sq(a, pdom, map, block_dim);
+    REAL r = L0Dev_Norm2Sq(a, pdom, map, block_dim);
     if (cpm.size > 1) {
         CPML2_AllReduce(&r, 1, MPI_DOUBLE, MPI_SUM, MPI_COMM_WORLD);
     }
     return r;
 }
 
-static double L2Dev_MaxDiag(Matrix<double> &a, Mapper &pdom, dim3 block_dim, CPMBase &cpm) {
+static REAL L2Dev_MaxDiag(Matrix<REAL> &a, Mapper &pdom, dim3 block_dim, CPMBase &cpm) {
     Mapper map(pdom, Gd);
-    double r = L0Dev_MaxDiag(a, pdom, map, block_dim);
+    REAL r = L0Dev_MaxDiag(a, pdom, map, block_dim);
     if (cpm.size > 1) {
         CPML2_AllReduce(&r, 1, MPI_DOUBLE, MPI_MAX, MPI_COMM_WORLD);
     }
